[refactor] Apply fixtures : categories
